requirements.txt
cffi==1.3.1
msgpack-python==0.4.6
